Move `scripts.schema` to `pycroft.model.alembic` and re-use
